clean views.py hopefully complete
