+++ myplugin/content/eduvmstore/views.py
@@ -9,7 +9,6 @@ from django.http import JsonResponse
 from django.shortcuts import render, redirect
 from django.urls import reverse_lazy
 from horizon import tabs, exceptions
-from openstack_dashboard import api
 from openstack_dashboard.api import glance, nova, cinder, keystone, neutron
 from django.views import generic
 from django.utils.translation import gettext_lazy as _
@@ -33,7 +32,7 @@ from django.shortcuts import get_object_or_404
 from django.views import View
 import base64
 import re
-from myplugin.content.eduvmstore.utils import get_token_id, search_app_templates, fetch_favorite_app_templates, get_images_data, get_image_data
+from myplugin.content.eduvmstore.utils import get_token_id, search_app_templates, fetch_favorite_app_templates, get_images_data, get_image_data, get_app_template, generate_pdf, generate_ssh_instructions_pdf, generate_cloud_config, generate_indented_content
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -163,7 +162,7 @@ class DetailsPageView(generic.TemplateView):
         :return: Context dictionary with AppTemplate and image details.
         """
         context = super().get_context_data(**kwargs)
-        app_template = self.get_app_template()
+        app_template = get_app_template(self.request, self.kwargs['template_id'])
         image_data = get_image_data(self.request, app_template.get('image_id', ''))
         created_at = app_template.get('created_at', '').split('T')[0]
 
@@ -194,35 +193,6 @@ class DetailsPageView(generic.TemplateView):
             return user.name
         except Exception:
             return user_id
-
-
-
-    def get_app_template(self):
-        """
-        Fetch a specific AppTemplate from the external database using token authentication.
-        :return: JSON response of AppTemplate details if successful, otherwise an empty dict.
-        """
-        token_id = get_token_id(self.request)
-        headers = {"X-Auth-Token": token_id}
-
-        try:
-            response = requests.get(
-                API_ENDPOINTS['app_template_detail'].format(template_id=self.kwargs['template_id']),
-                headers=headers,
-                timeout=10
-            )
-            response.raise_for_status()
-            app_template = response.json()
-
-
-            app_template.setdefault('instantiation_attributes', [])
-            app_template.setdefault('account_attributes', [])
-
-            return app_template
-
-        except requests.RequestException as e:
-            logging.error("Unable to retrieve app template details: %s", e)
-            return {"instantiation_attributes": [], "account_attributes": []}
 
 
 
@@ -384,7 +354,7 @@ class AppTemplateView(generic.TemplateView):
         template_name = self.request.GET.get("template")
 
         if template_id:
-            app_template = self.get_app_template(template_id)
+            app_template = get_app_template(self.request,template_id)
             image_data = get_image_data(self.request, app_template.get('image_id', ''))
             db_security_groups = [sg['name'] for sg in app_template.get('security_groups', [])]
         elif template_name in preset_examples and self.mode != "edit":
@@ -423,33 +393,6 @@ class AppTemplateView(generic.TemplateView):
 
         return context
 
-    def get_app_template(self, template_id):
-        """
-        Fetch a specific AppTemplate from the external database using token authentication.
-
-        This function retrieves the token ID from the request, constructs the headers,
-        and makes a GET request to the external API to fetch the AppTemplate details.
-        If the request is successful, it returns the JSON response. In case of an error,
-        it logs the error and returns an empty dictionary.
-
-        :param template_id: The ID of the AppTemplate to retrieve.
-        :type template_id: str
-        :return: JSON response of AppTemplate details if successful, otherwise an empty dict.
-        :rtype: dict
-        """
-        token_id = get_token_id(self.request)
-        headers = {"X-Auth-Token": token_id}
-        try:
-            response = requests.get(
-                API_ENDPOINTS['app_template_detail'].format(template_id=template_id),
-                headers=headers,
-                timeout=10
-            )
-            response.raise_for_status()
-            return response.json()
-        except requests.RequestException as e:
-            logging.error("Unable to retrieve app template details: %s", e)
-            return {}
 
 
     def get_security_groups(self):
@@ -464,214 +407,6 @@ class AppTemplateView(generic.TemplateView):
         except Exception as e:
             logging.error(f"Unable to retrieve security groups: {e}")
             return []
-
-
-def generate_pdf(accounts, name, app_template, created, instantiations, ip_address):
-    """
-    Generate a well-formatted PDF document containing user account information in a table format.
-
-    :param accounts: A list of dictionaries, where each dictionary contains user account details.
-    :type accounts: list
-    :param name: The name of the created instance.
-    :type name: str
-    :return: PDF-Datei als Bytes (nicht als HttpResponse!)
-    :rtype: bytes
-    """
-    buffer = io.BytesIO()
-    doc = SimpleDocTemplate(buffer, pagesize=letter)
-    elements = []
-    styles = getSampleStyleSheet()
-
-    title = Paragraph(f"<b>{name}</b>", styles['Title'])
-    elements.append(title)
-    elements.append(Spacer(1, 0.2 * inch))
-
-    subtitle = Paragraph(
-        f"Instantiation Attributes for the created instance {name} from the EduVMStore. "
-        f"This instance was created with the app template {app_template} on {created}.",
-        styles['Normal']
-
-    )
-    elements.append(subtitle)
-    if ip_address:
-        ip = Paragraph(f"IP Address: {ip_address}", styles['Normal'])
-        elements.append(ip)
-    elements.append(Spacer(1, 0.2 * inch))
-
-    if accounts:
-        elements.append(Paragraph("<b>Account Attributes</b>", styles['Heading2']))
-        elements.append(Spacer(1, 0.1 * inch))
-        all_keys = list(accounts[0].keys())
-        table_data = [all_keys]
-        for account in accounts:
-            row_values = [account.get(key, "N/A") for key in all_keys]
-            table_data.append(row_values)
-        table = Table(table_data, repeatRows=1)
-        table.setStyle(TableStyle([
-            ('BACKGROUND', (0, 0), (-1, 0), colors.grey),
-            ('TEXTCOLOR', (0, 0), (-1, 0), colors.whitesmoke),
-            ('ALIGN', (0, 0), (-1, -1), 'CENTER'),
-            ('FONTNAME', (0, 0), (-1, 0), 'Helvetica-Bold'),
-            ('BOTTOMPADDING', (0, 0), (-1, 0), 8),
-            ('BACKGROUND', (0, 1), (-1, -1), colors.white),
-            ('GRID', (0, 0), (-1, -1), 1, colors.black),
-        ]))
-        elements.append(table)
-        elements.append(Spacer(1, 0.2 * inch))
-
-
-    if instantiations:
-        elements.append(Paragraph("<b>Instantiation Attributes</b>", styles['Heading2']))
-        elements.append(Spacer(1, 0.1 * inch))
-
-        keys = list(instantiations[0].keys())
-        table_data_instantiation = []
-
-        header_row = ["Attributes"] + ["Values"]
-        table_data_instantiation.append(header_row)
-
-        for key in keys:
-            row = [key]
-            for inst in instantiations:
-                row.append(inst.get(key, "N/A"))
-            table_data_instantiation.append(row)
-
-        table_instantiation = Table(table_data_instantiation, repeatRows=1)
-        table_instantiation.setStyle(TableStyle([
-            ('BACKGROUND', (0, 0), (-1, 0), colors.grey),
-            ('TEXTCOLOR', (0, 0), (-1, 0), colors.whitesmoke),
-            ('ALIGN', (0, 0), (-1, -1), 'CENTER'),
-            ('FONTNAME', (0, 0), (-1, 0), 'Helvetica-Bold'),
-            ('BOTTOMPADDING', (0, 0), (-1, 0), 8),
-            ('BACKGROUND', (0, 1), (-1, -1), colors.white),
-            ('GRID', (0, 0), (-1, -1), 1, colors.black),
-        ]))
-        elements.append(table_instantiation)
-
-
-    doc.build(elements)
-    buffer.seek(0)
-
-    return buffer.getvalue()
-
-def generate_ssh_instructions_pdf(instances):
-    """
-    Generate a PDF containing SSH copy instructions for the user keys.
-
-    :param instances: A list of instance dictionaries with 'name', 'ip', and 'key'.
-    :type instances: list
-    :return: PDF content as bytes.
-    :rtype: bytes
-    """
-    buffer = io.BytesIO()
-    doc = SimpleDocTemplate(buffer, pagesize=letter)
-    elements = []
-    styles = getSampleStyleSheet()
-
-    title = Paragraph("<b>SSH Instructions for Downloading User Keys</b>", styles['Title'])
-    elements.append(title)
-    elements.append(Spacer(1, 0.2 * inch))
-    subtitle = Paragraph(
-        f"These instructions will help you to download the private ssh user keys from the instances."
-        f" Please use the following commands to copy the keys from the instances to your local machine.",
-        styles['Normal']
-
-    )
-    elements.append(subtitle)
-    elements.append(Spacer(1, 0.2 * inch))
-
-    for idx, instance in enumerate(instances, start=1):
-        name = instance.get('name', 'Unknown')
-        ip = instance.get('ip', 'Unknown')
-        key = instance.get('key', 'Unknown')
-        command = f"scp -i {key} -r ubuntu@{ip}:/home/ubuntu/user_keys/ ."
-
-        elements.append(Paragraph(f"<b>{name}</b>", styles['Heading2']))
-        elements.append(Spacer(1, 0.1 * inch))
-        elements.append(Paragraph(command, styles['Code']))
-        elements.append(Spacer(1, 0.2 * inch))
-
-    doc.build(elements)
-    buffer.seek(0)
-
-    return buffer.getvalue()
-
-
-
-def generate_cloud_config(accounts, backend_script, instantiations):
-    """
-    Generate a cloud-config file for user account creation and backend script execution.
-
-    :param accounts: A list of dictionaries with user account details.
-    :param backend_script: Backend script to be included in the cloud-config.
-    :param instantiations: A list of dictionaries with instantiation data.
-    :return: A string representing the complete cloud-config file.
-    """
-
-    users_content = ""
-    instantiations_content = ""
-
-    if accounts:
-        sorted_keys = list(accounts[0].keys())
-        users_content = "\n".join(
-            [":".join([account.get(key, "N/A") for key in sorted_keys]) for account in accounts]
-        )
-
-    if instantiations:
-        sorted_keys_instantiation = list(instantiations[0].keys())
-        instantiations_content = "\n".join(
-            [":".join([inst.get(key, "N/A") for key in sorted_keys_instantiation])
-             for inst in instantiations]
-        )
-
-    write_files_block = ""
-
-    if users_content:
-        write_files_block += f"""
-  - path: /etc/users.txt
-    content: |
-{generate_indented_content(users_content, indent_level=6)}
-    permissions: '0644'
-    owner: root:root
-"""
-
-    if instantiations_content:
-        write_files_block += f"""
-  - path: /etc/attributes.txt
-    content: |
-{generate_indented_content(instantiations_content, indent_level=6)}
-    permissions: '0644'
-    owner: root:root
-"""
-
-    cloud_config = "#cloud-config\n"
-    if write_files_block:
-        cloud_config += f"write_files:{write_files_block}"
-
-    if backend_script:
-        cloud_config += f"\n\n{backend_script}"
-
-    return cloud_config
-
-
-
-def generate_indented_content(content, indent_level=6):
-    """
-    Indent each line of the given content by a specified number of spaces.
-
-    This function takes a multi-line string and indents each line by a specified number of spaces.
-    It is useful for formatting content that needs to be indented consistently.
-
-    :param content: The multi-line string to be indented.
-    :type content: str
-    :param indent_level: The number of spaces to indent each line.
-    :type indent_level: int
-    :return: The indented multi-line string.
-    :rtype: str
-    """
-
-    indent = " " * indent_level
-    return "\n".join([indent + line for line in content.split("\n")])
 
 
 class InstancesView(generic.TemplateView):
@@ -706,7 +441,7 @@ class InstancesView(generic.TemplateView):
         try:
             num_instances = int(request.POST.get('instance_count', 1))
             base_name = request.POST.get('instances_name')
-            app_template = self.get_app_template()
+            app_template = get_app_template(self.request, self.kwargs['image_id'])
             image_id = app_template.get('image_id')
             script = app_template.get('script')
             app_template_name = app_template.get('name')
@@ -947,7 +682,7 @@ class InstancesView(generic.TemplateView):
 
     def get_network_name_by_id(self, request, network_id):
         try:
-            networks = api.neutron.network_list(request)
+            networks = neutron.network_list(request)
             for network in networks:
                 if network.id == network_id:
                     return network.name
@@ -1033,7 +768,8 @@ class InstancesView(generic.TemplateView):
         """
         context = super().get_context_data(**kwargs)
         app_template_id = self.kwargs['image_id']
-        app_template = self.get_app_template()
+        app_template = get_app_template(self.request, self.kwargs['image_id'])
+
 
         # Fetch available flavors from Nova
         context['flavors'] = self.get_flavors(app_template)
@@ -1076,7 +812,7 @@ class InstancesView(generic.TemplateView):
         :rtype: dict
         """
         try:
-            flavors = api.nova.flavor_list(self.request)
+            flavors = nova.flavor_list(self.request)
             if not flavors:
                 logging.error("No flavors returned from Nova API.")
                 return {}
@@ -1122,7 +858,8 @@ class InstancesView(generic.TemplateView):
         :return: A list of expected field names for account creation.
         :rtype: list
         """
-        app_template = self.get_app_template()
+        app_template = get_app_template(self.request, self.kwargs['image_id'])
+
 
         account_attributes = app_template.get('account_attributes')
 
@@ -1170,7 +907,7 @@ class InstancesView(generic.TemplateView):
         :return: A list of expected field names for account creation.
         :rtype: list
         """
-        app_template = self.get_app_template()
+        app_template = get_app_template(self.request, self.kwargs['image_id'])
 
         instantiation_attributes = app_template.get('instantiation_attributes')
 
@@ -1222,33 +959,13 @@ class InstancesView(generic.TemplateView):
         """
         try:
             tenant_id = self.request.user.tenant_id
-            networks = api.neutron.network_list_for_tenant(self.request, tenant_id)
+            networks = neutron.network_list_for_tenant(self.request, tenant_id)
             return {network.id: network.name for network in networks}
         except Exception as e:
             logging.error(f"Unable to fetch networks: {e}")
             return {}
 
-    # Get AppTemplate Details to display while launching an instance
-    def get_app_template(self):
-        """
-            Fetch a specific AppTemplate from the external database using token authentication.
-            :param token_id: Authentication token for API access.
-            :return: JSON response of AppTemplate details if successful, otherwise an empty dict.
-            :rtype: dict
-        """
-        token_id = get_token_id(self.request)
-        headers = {"X-Auth-Token": token_id}
 
-        try:
-            response = (requests.get(API_ENDPOINTS['app_template_detail'].format(
-                template_id=self.kwargs['image_id']),
-                headers=headers, timeout=10))
-
-            response.raise_for_status()
-            return response.json()
-        except requests.RequestException as e:
-            logging.error("Unable to retrieve app template details: %s", e)
-            return {}
 
     def format_description(self, description):
         """
